Removed some potentially un-needed code and debug
so this code works for me right now, the only problem which I don't mind, is when the location is first spawned in only one of the item stands is changed to allow the customizable SacrificeItem. It only works for all item stands when the location is loaded out of game scene and back in, for example dragonqueen's location 3 item stands. not really a huge problem but if it's something you can apply a quick fix for might aswell.

Other than that, if you could review the changes I've made and remove all unnecessary debug/code and only leave what's needed to work, I can merge and submit to nexus/thunderstore.